requirements.txt
Flask
openai
helius-sdk
solana
gunicorn